corregida estructura del proyecto
