+++ src/App.tsx
@@ -1,18 +1,15 @@
-import React from "react";
-import ResidentDashboard from "./pages/ResidenteDashboard";
-import Home from "./components/login/Home";
-import Authorizations from "./pages/Authorizations";
-import VisitHistory from "./pages/VisitHistory";
-import Settings from "./pages/Settings";
 import { Route, Routes } from "react-router-dom";
+import Dashboard from "./pages/Dashboard";
+import Authorizations from "./pages/Authorizations";
+import Settings from "./pages/Settings";
 import VisitEntryForm from "./components/visitRegistry/VisitEntryForm";
+import Home from "./components/login/Home";
 
 const App = () => {
   return (
     <Routes>
-      <Route path="/" element={<VisitEntryForm />} />
+      <Route path="/" element={<Dashboard />} />
       <Route path="/authorizations" element={<Authorizations />} />
-      <Route path="/visit-history" element={<VisitHistory />} />
       <Route path="/settings" element={<Settings />} />
     </Routes>
   );
