+++ src/components/visits/Header.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles/residente.module.css";
+import styles from "../../styles/visits.module.css";
 
 const Header = () => {
   return (
