+++ src/components/visits/QuickAction.tsx
@@ -1,5 +1,5 @@
 import { FaEdit, FaFileExport, FaClipboardList } from "react-icons/fa";
-import styles from "../styles/residente.module.css";
+import styles from "../../styles/visits.module.css";
 
 const QuickActions = () => {
   return (
