--- src/pages/ResidenteDashboard.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import Dashboard from "../components/residente/Dashboard";
-
-const ResidentDashboard = () => {
-  return <Dashboard />;
-};
-
-export default ResidentDashboard;
