+++ src/pages/Settings.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import styles from "../components/styles/residente.module.css";
+import styles from "../styles/visits.module.css";
 
 const Settings = () => {
   return (
