+++ src/routes/VisitRoutes.tsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from "react-router-dom";
-import Dashboard from "../components/residente/Dashboard";
+import Dashboard from "../pages/Dashboard";
 
-export default function ResidentRoutes() {
+export default function VisitsRoutes() {
   return (
     <Routes>
       <Route path="/dashboard" element={<Dashboard />} />
