+++ src/styles/animations.module.css
